+++ src/components/head-meta.astro
@@ -1,0 +1,17 @@
+---
+import { ClientRouter } from 'astro:transitions'
+import '../styles/global.css'
+---
+
+<head>
+  <meta charset="utf-8" />
+  <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
+  <meta name="viewport" content="width=device-width" />
+  <meta name="generator" content={Astro.generator} />
+  <link
+    href="https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,300;0,500;1,300&display=swap"
+    rel="stylesheet"
+  />
+  <title>Comics</title>
+  <ClientRouter />
+</head>
