+++ src/pages/[...week].astro
@@ -1,7 +1,7 @@
 ---
-import Page from '../Layouts/Page.astro'
+import Week from '../Layouts/week.astro'
 
 const { week } = Astro.params
 ---
 
-<Page {week} />
+<Week {week} />
