+++ src/pages/comic/[id].astro
@@ -1,68 +1,47 @@
 ---
+import Details from '../../Layouts/details.astro'
 import { getComicDetails } from '../../data/get-comic-details'
-import Footer from '../../components/footer.astro'
-import Nav from '../../components/nav.astro'
 import ComicDetails from '../../components/comic-details.astro'
 import { getNewComicBookDay } from '../../data/get-new-comic-book-day'
-import '../../styles/global.css'
 
-// Get the current week for nav
 const { currentWeek } = getNewComicBookDay()
 
-// Get the comic ID from the URL
 const { id } = Astro.params
 const comicDetails = await getComicDetails(id as string)
 ---
 
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
-    <meta name="viewport" content="width=device-width" />
-    <meta name="generator" content={Astro.generator} />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,300;0,500;1,300&display=swap"
-      rel="stylesheet"
-    />
-    <title>{comicDetails.issue || 'Comic Details'}</title>
-  </head>
-  <body class="w-screen bg-sky-700">
-    <Nav week={currentWeek} />
-    <div class="bg-slate-900">
-      <main class="mx-auto max-w-7xl px-4 py-8 sm:px-6 lg:px-8">
-        {
-          !comicDetails || Object.keys(comicDetails).length === 0 ? (
-            <div class="bg-slate-200 rounded-lg shadow-lg p-8 text-center">
-              <h1 class="text-2xl font-bold font-montserrat text-slate-900 mb-4">
-                Comic Not Found
-              </h1>
-              <p class="text-slate-700 mb-6">
-                Sorry, we couldn't find details for this comic. It may have been
-                removed or the ID is incorrect.
-              </p>
-              <a
-                href={`/${currentWeek}`}
-                class="inline-block px-4 py-2 bg-sky-700 text-white rounded-md hover:bg-sky-800 transition-colors"
-              >
-                Back to Comics List
-              </a>
-            </div>
-          ) : (
-            <div>
-              <ComicDetails comic={comicDetails} />
-              <div class="mt-6 text-center">
-                <a
-                  href={`/${currentWeek}`}
-                  class="inline-block px-4 py-2 bg-sky-700 text-white rounded-md hover:bg-sky-800 transition-colors"
-                >
-                  Back to Comics List
-                </a>
-              </div>
-            </div>
-          )
-        }
-      </main>
-    </div>
-    <Footer />
-  </body>
-</html>
+<Details>
+  <main class="mx-auto max-w-7xl px-4 py-8 sm:px-6 lg:px-8">
+    {
+      !comicDetails || Object.keys(comicDetails).length === 0 ? (
+        <div class="rounded-lg bg-slate-200 p-8 text-center shadow-lg">
+          <h1 class="font-montserrat mb-4 text-2xl font-bold text-slate-900">
+            Comic Not Found
+          </h1>
+          <p class="mb-6 text-slate-700">
+            Sorry, we couldn't find details for this comic. It may have been
+            removed or the ID is incorrect.
+          </p>
+          <a
+            href={`/${currentWeek}`}
+            class="inline-block rounded-md bg-sky-700 px-4 py-2 text-white transition-colors hover:bg-sky-800"
+          >
+            Back to Comics List
+          </a>
+        </div>
+      ) : (
+        <div>
+          <ComicDetails comic={comicDetails} />
+          <div class="mt-6 text-center">
+            <a
+              href={`/${currentWeek}`}
+              class="inline-block rounded-md bg-sky-700 px-4 py-2 text-white transition-colors hover:bg-sky-800"
+            >
+              Back to Comics List
+            </a>
+          </div>
+        </div>
+      )
+    }
+  </main>
+</Details>
